refactor: Refatorado código | desacoplado typedefs
