+++ src/Schema/TypeDefs/mutations.js
@@ -1,0 +1,9 @@
+const { gql } = require("apollo-server-express");
+
+const mutationTypeDefs = gql`
+  type Mutation {
+    createUser(name: String!, email: String!): User!
+  }
+`;
+
+module.exports = { mutationTypeDefs };
