+++ src/Schema/TypeDefs/subscriptions.js
@@ -1,0 +1,9 @@
+const { gql } = require("apollo-server-express");
+
+const subscriptionTypeDefs = gql`
+  type Subscription {
+    userAdded: User!
+  }
+`;
+
+module.exports = { subscriptionTypeDefs };
